@@ -7,11 +7,12 @@
     "email": "ryaneogden@gmail.com"
   },
   "scripts": {
-    "serve": "node cli/index.js -c tailwind.config.sample.js & vue-cli-service serve",
-    "build": "vue-cli-service build",
+    "dev": "vite",
+    "build": "vite build",
+    "serve": "vite preview",
     "build:docs": "npm run build && node ./cli export ./docs -c tailwind.config.sample.js",
     "prepublishOnly": "npm run lint --no-fix && npm run build",
-    "lint": "vue-cli-service lint",
+    "lint": "eslint --ext .js,.vue --ignore-path .gitignore --fix src",
     "test": "echo \"No test specified\"",
     "version": "npm run build:docs && git add docs"
   },
@@ -51,22 +52,16 @@
     "tailwindcss": "1 || 2 || 2.0.1-compat || 3"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^3.7.0",
-    "@vue/cli-plugin-eslint": "^3.7.0",
-    "@vue/cli-service": "^3.7.0",
+    "@vitejs/plugin-vue": "^5.0.3",
     "@vue/eslint-config-standard": "^4.0.0",
-    "babel-eslint": "^10.0.1",
-    "core-js": "^2.6.5",
     "defu": "^3.2.2",
     "eslint": "^5.16.0",
     "eslint-plugin-vue": "^5.0.0",
     "sass": "^1.26.10",
-    "sass-loader": "^9.0.3",
     "tailwindcss": "^1.4.4",
     "tailwindcss-dark-mode": "^1.1.6",
-    "vue": "^2.6.10",
-    "vue-draggable-resizable": "^2.2.0",
-    "vue-intersect": "^1.1.6",
-    "vue-template-compiler": "^2.5.21"
+    "vite": "^5.0.12",
+    "vue": "^3.4.15",
+    "vue-draggable-resizable": "^2.2.0"
   }
 }
